Replaced CentOS packages
The dockerized version of CentOS distribution from docker hub which is used as a starting point for out image build process is based on a "coreutils-single" GNU packages collection. That package conflicts with a regular "coreutils.rpm" which required by many other higher-level packages.

This CentOS packer profile replaced all packages with the standard ones.

centos8-replace-packages-all

The dockerized version of CentOS distribution from docker hub which is used as a starting point for out image build process is based on a "coreutils-single" GNU packages collection. That package conflicts with a regular "coreutils.rpm" which required by many other higher-level packages.

This CentOS packer profile replaced all packages with the standard ones.